sessions_mount_dir/djx6bg6ycfbcv2w07yftbf0uc/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
prometheus-client==0.19.0
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2
email-validator==2.1.0
python-dotenv==1.0.0